Module parts have IDs. WARN: ACTION REQUIRED
They aren't available for every module part unfortunately, so we can't
use them as IDs. However, we should still store them. Also, the module
ID is not enough as a foreign key without the module version.

This changes the table structure, so there is

ACTION REQUIRED: If you ran a version before this commit, please
manually delete the module_parts table (there should not be any
information stored in there anyways). On the next run, the table should
then be recreated with the additional columns.